refactors dungeon scene
